+++ hypothesis-python/docs/changes.rst
@@ -10,6 +10,23 @@ on `PyPI <https://pypi.org/project/hypothesis/>`__.
 Hypothesis 5.x
 ==============
 
+.. _v5.30.1:
+
+-------------------
+5.30.1 - 2020-09-04
+-------------------
+
+This patch makes the :doc:`ghostwriter <ghostwriter>` much more robust when
+passed unusual modules.
+
+- improved support for non-resolvable type annotations
+- :func:`~hypothesis.extra.ghostwriter.magic` can now write
+  :func:`~hypothesis.extra.ghostwriter.equivalent` tests
+- running :func:`~hypothesis.extra.ghostwriter.magic` on modules where some
+  names in ``__all__`` are undefined skips such names, instead of raising an error
+- :func:`~hypothesis.extra.ghostwriter.magic` now knows to skip mocks
+- improved handling of import-time errors found by the ghostwriter CLI
+
 .. _v5.30.0:
 
 -------------------